+++ tests/raftstore/mod.rs
@@ -34,4 +34,3 @@ mod test_snap;
 mod test_region_heartbeat;
 mod test_stale_peer;
 mod test_lease_read;
-mod test_large_entry;
